Split out the commodity_network.py File
